Add files via upload
The objective of this project is to help finding the best place for a new well for XYZ mining company.

Steps to choose the location:

Collect the oil well parameters in the selected region: oil quality and volume of reserves;
Build a model for predicting the volume of reserves in the new wells;
Pick the oil wells with the highest estimated values;
Pick the region with the highest total profit for the selected oil wells. We have data on oil samples from three regions. Parameters of each oil well in the region are already known. We are build a model that will help to pick the region with the highest profit margin.