Add files via upload
Project Background and Objectives
The sweet Lift Taxi company has collected historical data on taxi orders at airports. To attract more drivers during peak hours, we need to predict the number of taxi orders for the next hour. So we are building a model for such a prediction.

The RMSE metric on the test set should not be more than 48.